hours now in float fixed permission bug list only projects with versions order project by creation dat list only projects the user has permissions in list only open projects include only users that hvae permissions in the project fixed translation bug implemented input validation (negative numbers are converted into positive) refactored overall code

+++ db/migrate/001_add_shift_hours_to_user_preferences.rb
@@ -1,6 +1,6 @@
 class AddShiftHoursToUserPreferences < ActiveRecord::Migration
   def up
-    add_column :user_preferences, :shift_hours, :integer
+    add_column :user_preferences, :shift_hours, :float
   end
 
   def down
